backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg[binary]>=3.3.2
psycopg2-binary>=2.9.9
requests>=2.32.3
djangorestframework-simplejwt==5.3.1
argon2-cffi==23.1.0
uvicorn==0.30.1
ruff==0.4.8
python-dotenv==1.0.1


